Improve the speed of `IdentitySlerp` 3x by using pow.
This improves the speed of the IdentitySlerp by 3x,
from 966ns on average to 389ns now. This has been
achieved by using `Pow` internally, and unlike
v1.4.0 considers the case where the real component
is negative, and returns the same result as before.

Add a field to Quaternion root `_TO_STRING_CHAR`
which is a `number?`, and allows control over
the default `ToString` behaviour and number of
decimal points printed out. Setting to nil
will print out all decimal points, setting to any
number will print out that maximum amount of decimal
points for all components of the Quaternion.
Recommended value is 5.

Add license notice to the top of all src files.